# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TravelX API
  version: 1.3.0
  description: |
    # TravelX — 스마트 환전 예약 플랫폼 API 명세서

    **PRD v1.3 기준** (FPT IS Internship Project)

    여행자가 환율을 비교하고, 원하는 통화를 예약한 뒤, 선택한 지점에서 수령하는
    온라인 환전 예약 플랫폼의 REST API 명세입니다.

    ## 인증
    - 사용자 인증은 **Google OAuth2 로그인만** 지원하며, 로그인 후 발급되는 **JWT**를
      `Authorization: Bearer {token}` 헤더로 전달합니다.
    - 역할(Role)은 `USER` / `ADMIN` 두 가지입니다. 지점 직원(Staff) 기능(우대율 설정, QR 검증)은
      MVP에서 `ADMIN` 권한으로 수행하며, 추후 `BRANCH` 역할 분리를 검토합니다.

    ## 공통 에러 응답
    모든 에러는 아래 형식으로 반환됩니다.
    ```json
    {
      "code": "TIME_SLOT_FULL",
      "message": "선택한 픽업 시간대의 예약 정원이 가득 찼습니다.",
      "timestamp": "2026-07-13T10:30:00Z"
    }
    ```

    | HTTP | code 예시 | 상황 |
    |------|-----------|------|
    | 400 | `VALIDATION_ERROR` | 요청 값 형식 오류 |
    | 400 | `AMOUNT_LIMIT_EXCEEDED` | USD 10,000 상당액 이상 예약 시도 (PRD §19.1, §21.2) |
    | 400 | `AMOUNT_BELOW_MINIMUM` | 최소 환전 금액(10,000 VND 상당) 미만 (PRD §19.1) |
    | 401 | `AUTH_INVALID_TOKEN` | JWT 누락/만료/위조 |
    | 403 | `FORBIDDEN_ROLE` | 권한 없는 리소스 접근 |
    | 403 | `PHONE_NOT_VERIFIED` | 휴대폰 미인증 상태로 예약 시도 (KYC 1단계, PRD §21.3) |
    | 403 | `NO_SHOW_LIMIT` | 노쇼 이력에 따른 예약 제한 (PRD §19.2) |
    | 404 | `RESOURCE_NOT_FOUND` | 존재하지 않는 리소스 |
    | 409 | `STOCK_EXCEEDED` | 예약 전용 재고 부족 (PRD §20.1) |
    | 409 | `TIME_SLOT_FULL` | 30분 슬롯 정원 초과 (PRD §19.3) |
    | 409 | `ALREADY_CANCELLED` / `ALREADY_COMPLETED` | 취소/완료된 예약에 대한 중복 처리 |
...

    ## 예약 정책 요약 (PRD §19)
    - 보증금 없음, 방문 전 무료 취소
...
servers:
...
paths:
...
  /internal/ai/currencies/{code}/recommendations:
    post:
      tags: [AI Internal]
      operationId: internalCreateAiRecommendation
      summary: (Internal) AI 환율 추천 결과 저장
      description: |
        AI 모듈이 계산한 최신 환율 추천 등급 및 근거(AI_RECOMMENDATION)를 저장합니다.
        외부 접근이 차단된 Internal API입니다.
      security: []
      parameters:
        - $ref: '#/components/parameters/CurrencyCode'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiRecommendationCreateRequest'
      responses:
        '201':
          description: 저장 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 1
        '400':
          $ref: '#/components/responses/BadRequest'
  /internal/ai/currencies/{code}/signals:
    post:
      tags: [AI Internal]
      operationId: internalCreateAiSignal
      summary: (Internal) AI 통계 시그널 적재
      description: |
        AI 분석에 사용된 이동평균선 등 보조 지표 데이터를 저장합니다.
      security: []
      parameters:
        - $ref: '#/components/parameters/CurrencyCode'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiSignalCreateRequest'
      responses:
        '201':
          description: 저장 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 10
        '400':
          $ref: '#/components/responses/BadRequest'
  /internal/ai/currencies/{code}/backtests:
    post:
      tags: [AI Internal]
      operationId: internalCreateBacktestResult
      summary: (Internal) 백테스트 결과 저장
      description: |
        주기적으로 실행된 AI 모델 적중률 검증 결과를 저장합니다.
      security: []
      parameters:
        - $ref: '#/components/parameters/CurrencyCode'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BacktestCreateRequest'
      responses:
        '201':
          description: 저장 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 5
        '400':
          $ref: '#/components/responses/BadRequest'
  /internal/ai/macro-indicators:
    post:
      tags: [AI Internal]
      operationId: internalCreateMacroIndicator
      summary: (Internal) 거시경제 지표 적재
      description: |
        AI 모듈이 주기적으로 수집한 국가별 선행 지표(금리, 인플레이션 등)를 저장합니다.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MacroIndicatorCreateRequest'
      responses:
        '201':
          description: 저장 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 12
        '400':
          $ref: '#/components/responses/BadRequest'

  # ──────────────────────────── Branch Admin (Missing APIs) ────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Google OAuth2 로그인 후 발급된 JWT
  parameters:
    CurrencyCode:
      name: code
      in: path
      required: true
      description: ISO 4217 통화 코드
      schema:
        type: string
        example: USD
...
  responses:
    BadRequest:
      description: 'VALIDATION_ERROR — 요청 값 형식/범위 오류'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [result, code, message]
      properties:
        result:
          type: string
          enum: [FAIL]
        code:
          type: string
          description: 도메인 에러 코드 (info의 공통 에러 표 참조)
          example: TIME_SLOT_FULL
        message:
          type: string
          example: 선택한 픽업 시간대의 예약 정원이 가득 찼습니다.

    # ---------- Auth / User ----------
...
    AiRecommendationCreateRequest:
      type: object
      required: [recommendation, rationale, confidenceScore, modelVersion]
      properties:
        signalIds:
          type: array
          description: "이 추천 결과의 근거가 된 RECOMMENDATION_SIGNAL ID 목록 (다대다 매핑용)"
          items:
            type: integer
            format: int64
          example: [101, 102]
        recommendation:
          type: string
          enum: [INCREASING, NEUTRAL, DECREASING]
          example: INCREASING
        rationale:
          type: string
          example: "이동평균 20일선 돌파 및 단기 골든크로스 발생"
        confidenceScore:
          type: number
          format: double
          example: 0.85
        modelVersion:
          type: string
          example: "v1.2.0"
        expiresAt:
          type: string
          format: date-time
          example: "2026-07-22T00:00:00Z"
    AiSignalCreateRequest:
      type: object
      required: [signalType, windowDays, value]
      properties:
        signalType:
          type: string
          enum: [MA, RSI, MACD, STD]
          example: MA
        windowDays:
          type: integer
          example: 20
        value:
          type: number
          format: double
          example: 1380.5
    BacktestCreateRequest:
      type: object
      required: [currencyId, strategyType, periodStart, periodEnd, totalSignals, correctSignals, accuracyRate]
      properties:
        currencyId:
          type: integer
          format: int64
          example: 1
        strategyType:
          type: string
          enum: [LINEAR_REGRESSION, BASE_LINE, LOGISTIC_REGRESSION]
          example: LINEAR_REGRESSION
        periodStart:
          type: string
          format: date
          example: "2025-01-01"
        periodEnd:
          type: string
          format: date
          example: "2026-01-01"
        totalSignals:
          type: integer
          example: 100
        correctSignals:
          type: integer
          example: 65
        accuracyRate:
          type: number
          format: double
          example: 0.65
    MacroIndicatorCreateRequest:
      type: object
      required: [countryCode, indicatorType, value, recordedAt]
      properties:
        countryCode:
          type: string
          example: "US"
        indicatorType:
          type: string
          example: "CPI"
        value:
          type: number
          format: double
          example: 3.1
        recordedAt:
          type: string
          format: date
          example: "2026-07-21"
security:
  - bearerAuth: []